+++ js/markdown/package.json
@@ -26,10 +26,10 @@
 		"@gradio/icons": "workspace:^",
 		"@gradio/statustracker": "workspace:^",
 		"@gradio/utils": "workspace:^",
+		"@gradio/sanitize": "workspace:^",
 		"@types/dompurify": "^3.0.2",
 		"@types/katex": "^0.16.0",
 		"@types/prismjs": "1.26.4",
-		"amuchina": "^1.0.12",
 		"dom-parser": "^1.1.5",
 		"github-slugger": "^2.0.0",
 		"isomorphic-dompurify": "^2.14.0",
@@ -37,12 +37,10 @@
 		"marked": "^12.0.0",
 		"marked-gfm-heading-id": "^3.1.2",
 		"marked-highlight": "^2.0.1",
-		"prismjs": "1.29.0",
-		"sanitize-html": "^2.13.0"
+		"prismjs": "1.29.0"
 	},
 	"devDependencies": {
-		"@gradio/preview": "workspace:^",
-		"@types/sanitize-html": "^2.13.0"
+		"@gradio/preview": "workspace:^"
 	},
 	"peerDependencies": {
 		"svelte": "^4.0.0"
