+++ js/sanitize/index.d.ts
@@ -1,0 +1,1 @@
+export declare function sanitize(source: string, root: string): string;
